backend/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic
python-dotenv
httpx
numpy
scikit-learn
pymongo
difflib — difflib